requirements_prod.txt
fastapi==0.92.0
gunicorn==20.1.0
loguru==0.6.0
motor==3.1.1
python-dotenv==0.21.1
uvicorn[standard]==0.20.0
